requirements.txt
flask>=2.3.0
anthropic>=0.50.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
